+++ API-Plantacao/src/main/java/com/tiago/apiplantacao/repository/PlantaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.tiago.apiplantacao.model.Planta;
 
-public interface PlantaRepository extends JpaRepository<Planta, Integer>{
+public interface PlantaRepository extends JpaRepository<Planta, Long>{
 
 }
