+++ API-Plantacao/src/main/java/com/tiago/apiplantacao/repository/ResponsavelRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.tiago.apiplantacao.model.Responsavel;
 
-public interface ResponsavelRepository extends JpaRepository<Responsavel, Integer>{
+public interface ResponsavelRepository extends JpaRepository<Responsavel, Long>{
 
 }
